Combine e2e tests into a single comprehensive test suite
- Combined all e2e tests into integration.spec.ts
- Organized tests into logical groups (Extension Setup, UI and Interaction, History Sync)
- Fixed TypeScript error in history recording test
- Improved test organization and readability
